requirements.txt
# Generated from requirements/base.in using pip-compile. Do not edit by hand.
-c constraints.txt
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
yfinance==0.2.43
matplotlib==3.9.2
scipy==1.13.1
pyyaml==6.0.2
streamlit==1.39.0
plotly==5.24.1
jsonschema==4.23.0
